Change heading styles from h2 to h3
On the one hand, this is a hacky workaround to deal with the fact that the h3 appearance is what I want for the headings. (We really need to go through and improve the CSS site-wide.) But in a different light, this is actually what I want here, as these headings are less "big" headings are more intros into the content they directly precede.